refactor: Origin runtime settings into origin service desktop
