+++ ui/desktop/electron-app/src/models/runtime-settings.js
@@ -1,34 +1,7 @@
-const { net } = require('electron');
-
-const requestTimeoutSeconds = 10;
-// Simple promise wrapper around Electron's net.request feature.
-// https://www.electronjs.org/docs/api/client-request
-const netRequestPromise = (url) =>
-  new Promise((resolve, reject) => {
-    try {
-      const request = net.request(url);
-      // We don't worry about canceling the request timeout timer once a
-      // request ends because:  request.abort() has no effect after a request
-      // ends and the promise will already have been completed, so a rejection
-      // has no effect either.
-      const requestTimeout = setTimeout(() => {
-        request.abort();
-        reject(new Error('Request timeout'));
-      }, requestTimeoutSeconds * 1000);
-      request.on('error', reject);
-      request.on('response', resolve);
-      request.end();
-    } catch (e) {
-      reject(e);
-    }
-  });
-
-// Provides a singleton class instance to enable a consistent view of
-// runtime settings across the application.
+const netRequestPromise = require('../helpers/net-request-promise.js');
 
 // Runtime settings are any configuration, especially user-specified,
 // set up at runtime.  For example, the Boundary origin is a runtime setting.
-
 class RuntimeSettings {
   // Internal private origin is exposed via getter/setter below.
   #origin = undefined;
@@ -84,6 +57,4 @@ class RuntimeSettings {
   }
 }
 
-const runtimeSettings = new RuntimeSettings();
-
-module.exports = runtimeSettings;
+module.exports = RuntimeSettings;
